chore: organize the transfer data in a dto
