+++ internal/handlers/question.go
@@ -6,6 +6,7 @@ import (
 	"io"
 	"net/http"
 
+	"github.com/mbuchoff/hackathon_backend_230909/internal/dto"
 	"github.com/mbuchoff/hackathon_backend_230909/internal/services/translate"
 )
 
@@ -21,28 +22,28 @@ func AnswerQuestion(w http.ResponseWriter, r *http.Request) {
 		body, err := io.ReadAll(r.Body)
 		if err != nil {
 			w.WriteHeader(http.StatusInternalServerError)
-			json.NewEncoder(w).Encode(ResponseError{Message: "Internal Server Error"})
+			json.NewEncoder(w).Encode(dto.ResponseError{Message: "Internal Server Error"})
 			return
 		}
 		fmt.Println(string(body))
 		// extract the phrase from the request body and attach it to the payload and send to transtate the text
-		textToBeTranslated := Payload{Text: string(body)}
+		textToBeTranslated := dto.Payload{Text: string(body)}
 
 		// translate the phrase using our function translateTexts and return the translated phrase
 		translatedPhrase, err := translate.TranslateText(textToBeTranslated.Text)
 		if err != nil {
 			w.WriteHeader(http.StatusInternalServerError)
-			json.NewEncoder(w).Encode(ResponseError{Message: "Internal Server Error"})
+			json.NewEncoder(w).Encode(dto.ResponseError{Message: "Internal Server Error"})
 			return
 		}
 		fmt.Println("Translated phrase:", translatedPhrase)
 
 		// Write the response body
 		w.WriteHeader(http.StatusOK)
-		json.NewEncoder(w).Encode(Response{Message: translatedPhrase})
+		json.NewEncoder(w).Encode(dto.Response{Message: translatedPhrase})
 
 	} else {
 		w.WriteHeader(http.StatusMethodNotAllowed)
-		json.NewEncoder(w).Encode(ResponseError{Message: "Method not allowed"})
+		json.NewEncoder(w).Encode(dto.ResponseError{Message: "Method not allowed"})
 	}
 }
